+++ R/coerce-SummarizedExperiment.R
@@ -1,6 +1,8 @@
-#' Coerce to `SummarizedExperiment`
+#' Coerce to summarized experiment
 #'
-#' Improved method for reliably coercing objects that extend
+#' Coerce to `SummarizedExperiment`.
+#'
+#' Improved S3 methods for reliably coercing objects that extend
 #' `RangedSummarizedExperiment` to a standard `SummarizedExperiment`, that
 #' doesn't drop [`rowData()`][SummarizedExperiment::rowData].
 #'
@@ -9,11 +11,17 @@
 #' @inheritParams params
 #'
 #' @examples
+#' suppressPackageStartupMessages(library(SummarizedExperiment))
 #' data(rse, package = "acidtest")
 #'
-#' ## SummarizedExperiment ====
-#' x <- as.SummarizedExperiment(rse)
-#' x
+#' ## RangedSummarizedExperiment ===
+#' ## Current S4 coercion drops row metadata.
+#' names(metadata(rowRanges(rse)))
+#' se <- as(rse, "SummarizedExperiment")
+#' names(metadata(rowData(se)))
+#' ## Our alternate S3 method preserves the metadata.
+#' se <- as.SummarizedExperiment(rse)
+#' names(metadata(rowData(se)))
 NULL
 
 
@@ -28,26 +36,51 @@ as.SummarizedExperiment <-  # nolint
 
 
 
-# Note that our method here keeps track of `rowData()` when coercing an object
-# that extends RangedSummarizedExperiment to SummarizedExperiment. This bug
-# needs to be fixed in SummarizedExperiment.
-
-#' @method as.SummarizedExperiment default
+#' @rdname coerce-SummarizedExperiment
 #' @export
-as.SummarizedExperiment.default <-  # nolint
+as.SummarizedExperiment.RangedSummarizedExperiment <-  # nolint
     function(x) {
-        if (is(x, "RangedSummarizedExperiment")) {
-            rowMeta <- metadata(rowRanges(x))
-            x <- as(x, "RangedSummarizedExperiment")
-        } else {
-            rowMeta <- metadata(rowData(x))
-        }
-        x <- as(x, "SummarizedExperiment")
-        metadata(rowData(x)) <- rowMeta
-        x
+        `coerce,RangedSummarizedExperiment,SummarizedExperiment`(x)
+    }
+
+
+
+#' @rdname coerce-SummarizedExperiment
+#' @export
+as.SummarizedExperiment.SummarizedExperiment <-  # nolint
+    function(x) {
+        `coerce,SummarizedExperiment,SummarizedExperiment`(x)
     }
 
 
 
 # S4 ===========================================================================
-# S4 methods are already defined in SummarizedExperiment.
+# Can't set S4 coercion methods because RSE and SE aren't local, modifiable.
+
+# Coercion methods of interest:
+# - from="RangedSummarizedExperiment", to="SummarizedExperiment"
+# - from="SingleCellExperiment", to="RangedSummarizedExperiment"
+# - from="SingleCellExperiment", to="SummarizedExperiment"
+
+
+
+# Updated 2019-07-19.
+`coerce,RangedSummarizedExperiment,SummarizedExperiment` <-  # nolint
+    function(from) {
+        rowMeta <- metadata(rowRanges(from))
+        to <- as(from, "RangedSummarizedExperiment")
+        to <- as(from, "SummarizedExperiment")
+        metadata(rowData(to)) <- rowMeta
+        to
+    }
+
+
+
+# Updated 2019-07-19.
+`coerce,SummarizedExperiment,SummarizedExperiment` <-  # nolint
+    function(from) {
+        rowMeta <- metadata(rowData(from))
+        to <- as(from, "SummarizedExperiment")
+        metadata(rowData(to)) <- rowMeta
+        to
+    }