+++ R/coerce-data.table.R
@@ -1,6 +1,10 @@
-#' Coerce to `data.table`
+#' Coerce to data table
+#'
+#' Coerce to `data.table`.
 #'
 #' @name coerce-data.table
+#' @inheritParams data.table::as.data.table
+#' @inheritParams params
 #'
 #' @details
 #' Our defined methods attempt to improve on the defaults in the data.table
@@ -53,6 +57,7 @@ NULL
 #' @rdname coerce-data.table
 #' @name as.data.table
 #' @importFrom data.table as.data.table
+#' @usage as.data.table(x, keep.rownames = FALSE, ...)
 #' @export
 NULL
 
@@ -61,12 +66,13 @@ NULL
 # Note that we're matching `as_tibble()` convention here, using "rowname" as
 # column for row names assignment. We also using similar internal assert checks
 # here, allowing atomic and/or list columns only.
-# Updated 2019-07-11.
-#' @method as.data.table DataFrame
+
+#' @rdname coerce-data.table
 #' @export
+# Updated 2019-07-19.
 as.data.table.DataFrame <-  # nolint
     function(x, keep.rownames = TRUE, ...) {  # nolint
-        x <- .coerceDataFrame(x)
+        x <- `.coerce,DataFrame,data.frame`(x)
         if (!hasRownames(x)) {
             keep.rownames <- FALSE  # nolint
         }
@@ -77,9 +83,10 @@ as.data.table.DataFrame <-  # nolint
 
 # The default handling from data.frame isn't clean, so add this.
 # Default method will warn: `Arguments in '...' ignored`.
-# Updated 2019-07-11.
-#' @method as.data.table GRanges
+
+#' @rdname coerce-data.table
 #' @export
+# Updated 2019-07-19.
 as.data.table.GRanges <-  # nolint
     function(x, keep.rownames = TRUE, ...) {  # nolint
         x <- as(x, "data.frame")
@@ -91,60 +98,79 @@ as.data.table.GRanges <-  # nolint
 
 
 
-# Updated 2019-07-11.
-#' @method as.data.table IPosRanges
+#' @rdname coerce-data.table
 #' @export
+# Updated 2019-07-19.
 as.data.table.IPosRanges <- as.data.table.GRanges
 
 
 
 # S4 ===========================================================================
-# Updated 2019-07-11.
+# Updated 2019-07-19.
+`coerce,ANY,data.table` <-  # nolint
+    function(from) {
+        as.data.table(from)
+    }
+
+
+# Updated 2019-07-19.
+`coerce,data.frame,data.table` <-  # nolint
+    `coerce,ANY,data.table`
+
+
+
 #' @rdname coerce-data.table
 #' @name coerce,data.frame,data.table-method
 setAs(
     from = "data.frame",
     to = "data.table",
-    def = function(from) {
-        as.data.table(from)
-    }
+    def = `coerce,data.frame,data.table`
 )
 
 
 
-# Updated 2019-07-11.
+# Updated 2019-07-19.
+`coerce,DataFrame,data.table` <-  # nolint
+    `coerce,ANY,data.table`
+
+
+
 #' @rdname coerce-data.table
 #' @name coerce,DataFrame,data.table-method
 setAs(
     from = "DataFrame",
     to = "data.table",
-    def = function(from) {
-        as.data.table(from)
-    }
+    def = `coerce,DataFrame,data.table`
 )
 
 
 
-# Updated 2019-07-11.
+# Updated 2019-07-19.
+`coerce,GRanges,data.table` <-  # nolint
+    `coerce,ANY,data.table`
+
+
+
 #' @rdname coerce-data.table
 #' @name coerce,GRanges,data.table-method
 setAs(
     from = "GRanges",
     to = "data.table",
-    def = function(from) {
-        as.data.table(from)
-    }
+    def = `coerce,GRanges,data.table`
 )
 
 
 
-# Updated 2019-07-11.
+# Updated 2019-07-19.
+`coerce,IPosRanges,data.table` <-  # nolint
+    `coerce,ANY,data.table`
+
+
+
 #' @rdname coerce-data.table
 #' @name coerce,IPosRanges,data.table-method
 setAs(
     from = "IPosRanges",
     to = "data.table",
-    def = function(from) {
-        as.data.table(from)
-    }
+    def = `coerce,IPosRanges,data.table`
 )